Allow overwriting selected text when textbox full
If a text input element already contains the max number of characters as
defined by the "maxlength" RML attribute of the element, then if the
user attempts to type in another character, the attempt is correctly
ignored, since the textbox already contains maxlength characters.

However, if the user first selects at least one of the existing
characters in the text input element (e.g. by placing the cursor to the
right of a character in the text input element and pressing
shift + left arrow), then when the user attempts to type in another
character, the attempt should not be ignored.  The input character
should overwrite whatever block of characters are selected.  The
proposed change in this commit enables this behavior.